[EM-11] controller refactored in web package
