--- src/main/java/it/andmora/expensesmonitor/controller/PaymentController.java
@@ -1,9 +1,0 @@
-package it.andmora.expensesmonitor.controller;
-
-import it.andmora.expensesmonitor.controller.dto.PaymentDto;
-
-public interface PaymentController {
-
-  PaymentDto createPayment(PaymentDto paymentDto);
-
-}
