+++ src/main/java/it/andmora/expensesmonitor/web/PaymentController.java
@@ -1,0 +1,9 @@
+package it.andmora.expensesmonitor.web;
+
+import it.andmora.expensesmonitor.web.dto.PaymentDto;
+
+public interface PaymentController {
+
+  PaymentDto createPayment(PaymentDto paymentDto);
+
+}
